Move js files, start fixing #25
