+++ circle.yml
@@ -1,3 +1,4 @@
 machine:
   node:
     version: 0.10.28
+  pwd: $HOME/tools/js
