--- index.js
@@ -1,10 +1,0 @@
-module.exports = {
-  blocks: require('./enums/blocks'),
-  biomes: require('./enums/biomes'),
-  items: require('./enums/items'),
-  recipes: require('./enums/recipes'),
-  instruments: require('./enums/instruments'),
-  materials: require('./enums/materials'),
-  entities: require('./enums/entities'),
-  protocol: require('./enums/protocol')
-};
