--- package.json
@@ -1,22 +1,0 @@
-{
-  "name": "minecraft-data",
-  "version": "0.4.0",
-  "description": "provide minecraft data for minecraft clients, servers and libraries",
-  "main": "index.js",
-  "scripts": {
-    "test": "mocha --reporter spec"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git://github.com/PrismarineJS/minecraft-data.git"
-  },
-  "license": "MIT",
-  "devDependencies": {
-    "jsonschema": "~1.0.1",
-    "mocha": "~2.2.1",
-    "nodemw": "~0.4.2",
-    "underscore" : "~1.8.2",
-    "async":"~0.9.0",
-    "batch":"~0.5.2"
-  }
-}
