+++ tools/js/audit/audit_block_enum.js
@@ -1,4 +1,4 @@
-var blocks = require('../enums/blocks');
+var blocks = require('../../../enums/blocks');
 
 var all = [];
 for (var key in blocks) {
