+++ tools/js/burger_extractor/generate_enums.js
