+++ tools/js/decompiled_extractor/protocol_extractor.js
