+++ tools/js/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "minecraft-data-tools",
+  "private":"true",
+  "description": "minecraft-data tools",
+  "scripts": {
+    "test": "mocha --reporter spec"
+  },
+  "dependencies": {
+    "jsonschema": "~1.0.1",
+    "mocha": "~2.2.1",
+    "nodemw": "~0.4.2",
+    "underscore" : "~1.8.2",
+    "async":"~0.9.0",
+    "batch":"~0.5.2"
+  }
+}
