+++ tools/js/test/test.js
@@ -11,8 +11,8 @@ describe("minecraft-data", function() {
   this.timeout(60 * 1000);
     enums.forEach(function(enumName){
       it(enumName+".json is valid",function(){
-        var instance = require('../enums/'+enumName+'.json');
-        var schema = require('../enums_schemas/'+enumName+'_schema.json');
+        var instance = require('../../../enums/'+enumName+'.json');
+        var schema = require('../../../enums_schemas/'+enumName+'_schema.json');
         var result = v.validate(instance, schema);
         assert.strictEqual(result.errors.length,0,require('util').inspect(result.errors,{'depth':4}));
       })
