+++ tools/js/wiki_extractor/lib/find_item_object_by_name.js
