+++ tools/js/wiki_extractor/lib/id_table_template_parser.js
