+++ tools/js/wiki_extractor/test/test_block_extractor.js
