+++ tools/js/wiki_extractor/test/test_dvt_parser.js
