+++ tools/js/wiki_extractor/test/test_item_extractor.js
