+++ tools/js/wiki_extractor/test/test_protocol_extractor.js
